+++ walkability/components/categorise_paths/path_categorisation_artifacts.py
@@ -43,13 +43,14 @@ def build_path_categorisation_artifact(
 
     surface_quality = build_surface_quality_artifact(
         paths_line=paths_line,
+        paths_polygon=paths_polygon,
         walkable_categories=walkable_categories,
         resources=resources,
     )
 
-    smoothness = build_smoothness_artifact(paths_line=paths_line, resources=resources)
+    smoothness = build_smoothness_artifact(paths_line=paths_line, paths_polygon=paths_polygon, resources=resources)
 
-    surface_type = build_surface_artifact(paths_line=paths_line, resources=resources)
+    surface_type = build_surface_artifact(paths_line=paths_line, paths_polygon=paths_polygon, resources=resources)
 
     summary_artifacts = []
     if areal_summaries:
@@ -82,7 +83,7 @@ def build_walkable_paths_artifact(
 
     return create_geojson_artifact(
         features=walkable_locations.geometry,
-        layer_name='Walkable Path Category',
+        layer_name='Path Category',
         caption=Path('resources/components/categorise_paths/path_categorisation_caption.md').read_text(),
         description=Path('resources/components/categorise_paths/path_categorisation_description.md').read_text(),
         label=walkable_locations.category.apply(lambda r: r.value).to_list(),
@@ -106,35 +107,46 @@ def generate_detailed_pavement_quality_mapping_info() -> str:
 
 
 def build_surface_quality_artifact(
-    paths_line: gpd.GeoDataFrame, walkable_categories: Set[PathCategory], resources: ComputationResources
+    paths_line: gpd.GeoDataFrame,
+    paths_polygon: gpd.GeoDataFrame,
+    walkable_categories: Set[PathCategory],
+    resources: ComputationResources,
 ) -> Optional[_Artifact]:
     paths_line = next(subset_walkable_paths(paths_line, walkable_categories=walkable_categories))
-    if paths_line.empty:
+    surface_quality_locations = pd.concat([paths_line, paths_polygon], ignore_index=True)
+    if surface_quality_locations.empty:
         return
-    paths_line['color'] = generate_colors(paths_line.quality_rating, min=0.0, max=1.0, cmap_name='coolwarm_r')
+    surface_quality_locations['color'] = generate_colors(
+        surface_quality_locations.quality_rating, min=0.0, max=1.0, cmap_name='coolwarm_r'
+    )
     return create_geojson_artifact(
-        features=paths_line.geometry,
+        features=surface_quality_locations.geometry,
         layer_name='Surface Quality',
         caption=Path('resources/components/categorise_paths/surface_quality_caption.md').read_text(),
         description=Path('resources/components/categorise_paths/surface_quality_description.md').read_text()
         + generate_detailed_pavement_quality_mapping_info(),
-        label=paths_line.quality.apply(lambda r: r.value).to_list(),
-        color=paths_line.color.to_list(),
+        label=surface_quality_locations.quality.apply(lambda r: r.value).to_list(),
+        color=surface_quality_locations.color.to_list(),
         legend_data=get_surface_quality_legend(),
         resources=resources,
         filename='pavement_quality',
     )
 
 
-def build_smoothness_artifact(paths_line: gpd.GeoDataFrame, resources: ComputationResources) -> _Artifact:
-    paths_line['color'] = generate_colors(paths_line.smoothness_rating, cmap_name='coolwarm_r', min=0.0, max=1.0)
+def build_smoothness_artifact(
+    paths_line: gpd.GeoDataFrame, paths_polygon: gpd.GeoDataFrame, resources: ComputationResources
+) -> _Artifact:
+    smoothness_locations = pd.concat([paths_line, paths_polygon], ignore_index=True)
+    smoothness_locations['color'] = generate_colors(
+        smoothness_locations.smoothness_rating, cmap_name='coolwarm_r', min=0.0, max=1.0
+    )
     return create_geojson_artifact(
-        features=paths_line.geometry,
+        features=smoothness_locations.geometry,
         layer_name='Smoothness',
         caption=Path('resources/components/categorise_paths/smoothness_caption.md').read_text(),
         description=Path('resources/components/categorise_paths/smoothness_description.md').read_text(),
-        label=paths_line.smoothness.apply(lambda r: r.name).to_list(),
-        color=paths_line.color.to_list(),
+        label=smoothness_locations.smoothness.apply(lambda r: r.name).to_list(),
+        color=smoothness_locations.color.to_list(),
         legend_data=get_smoothness_legend(),
         resources=resources,
         filename='smoothness',
@@ -142,15 +154,18 @@ def build_smoothness_artifact(paths_line: gpd.GeoDataFrame, resources: Computati
     )
 
 
-def build_surface_artifact(paths_line: gpd.GeoDataFrame, resources: ComputationResources) -> _Artifact:
-    paths_line['color'] = generate_colors(paths_line.surface_rating, cmap_name='tab10', min=0.0, max=1.0)
+def build_surface_artifact(
+    paths_line: gpd.GeoDataFrame, paths_polygon: gpd.GeoDataFrame, resources: ComputationResources
+) -> _Artifact:
+    surface_locations = pd.concat([paths_line, paths_polygon], ignore_index=True)
+    surface_locations['color'] = generate_colors(surface_locations.surface_rating, cmap_name='tab10', min=0.0, max=1.0)
     return create_geojson_artifact(
-        features=paths_line.geometry,
+        features=surface_locations.geometry,
         layer_name='Surface Type',
         caption=Path('resources/components/categorise_paths/surface_caption.md').read_text(),
         description=Path('resources/components/categorise_paths/surface_description.md').read_text(),
-        label=paths_line.surface.apply(lambda r: r.name).to_list(),
-        color=paths_line.color.to_list(),
+        label=surface_locations.surface.apply(lambda r: r.name).to_list(),
+        color=surface_locations.color.to_list(),
         legend_data=get_surface_type_legend(),
         resources=resources,
         filename='surface_type',