+++ walkability/components/naturalness/naturalness_artifacts.py
@@ -1,4 +1,6 @@
 from pathlib import Path
+
+import pandas as pd
 from climatoology.base.baseoperator import _Artifact
 from climatoology.base.computation import ComputationResources
 from climatoology.base.artifact import (
@@ -15,12 +17,14 @@ from pydantic_extra_types.color import Color
 
 
 def build_naturalness_artifact(
-    naturalness: gpd.GeoDataFrame,
+    naturalness_line_paths: gpd.GeoDataFrame,
+    naturalness_polygon_paths: gpd.GeoDataFrame,
     resources: ComputationResources,
     cmap_name: str = 'YlGn',
 ) -> _Artifact:
+    naturalness_locations = pd.concat([naturalness_line_paths, naturalness_polygon_paths], ignore_index=True)
     # If no good data is returned (e.g. due to an error), return a text artifact with a simple message
-    if naturalness['naturalness'].isna().all():
+    if naturalness_locations['naturalness'].isna().all():
         text = 'There was an error calculating naturalness in this computation. Contact the developers for more information.'
         return create_markdown_artifact(
             text=text,
@@ -31,17 +35,17 @@ def build_naturalness_artifact(
         )
 
     # Set negative values (e.g. water) to 0, and set nan's to -999 to colour grey for unknown
-    naturalness['naturalness_col'] = naturalness['naturalness'].copy()
-    naturalness.loc[naturalness['naturalness'] < 0, 'naturalness_col'] = 0
-    naturalness.loc[naturalness['naturalness'].isna(), 'naturalness_col'] = -999
+    naturalness_locations['naturalness_col'] = naturalness_locations['naturalness'].copy()
+    naturalness_locations.loc[naturalness_locations['naturalness'] < 0, 'naturalness_col'] = 0
+    naturalness_locations.loc[naturalness_locations['naturalness'].isna(), 'naturalness_col'] = -999
 
     # Clean data for labels
-    naturalness['naturalness'] = naturalness['naturalness'].round(2)
+    naturalness_locations['naturalness'] = naturalness_locations['naturalness'].round(2)
 
     # Define colors and legend
     cmap = matplotlib.colormaps.get(cmap_name)
     cmap.set_under('#808080')
-    color = naturalness['naturalness_col'].apply(lambda v: Color(to_hex(cmap(v))))
+    color = naturalness_locations['naturalness_col'].apply(lambda v: Color(to_hex(cmap(v))))
     legend = ContinuousLegendData(
         cmap_name=cmap_name,
         ticks={'Low (0)': 0.0, 'High (1)': 1.0},
@@ -49,11 +53,11 @@ def build_naturalness_artifact(
 
     # Build artifact
     return create_geojson_artifact(
-        features=naturalness.geometry,
+        features=naturalness_locations.geometry,
         layer_name='Naturalness',
         caption=Path('resources/components/naturalness/caption.md').read_text(),
         description=Path('resources/components/naturalness/description.md').read_text(),
-        label=naturalness.naturalness.to_list(),
+        label=naturalness_locations.naturalness.to_list(),
         color=color,
         legend_data=legend,
         resources=resources,